Restructure of file
App -> ChatApp
New App renders Tabbar that routes to ChatApp and the other tabs
