+++ src/components/ChatApp/ChatApp.stories.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import ChatApp from './ChatApp';
+
+export default {
+  title: 'ChatApp',
+};
+
+export const Default = () => <ChatApp />;
